moved all tasks local into gulpfile
