+++ app/templates/_package.json
@@ -26,7 +26,6 @@
         "load-grunt-config": "~0.19.1",
         "time-grunt": "^1.3.0"<% } else if (taskRunner === 'gulp') {%>,
         "gulp": "~3.9.1",
-        "gulp-require-tasks": "~1.0.3",
         "run-sequence": "~1.1.5",
         "gulp-sass": "~2.2.0",
         "gulp-sourcemaps": "~1.5.2",
