--- app/templates/tasks/copy/_annotations.js
@@ -1,15 +1,0 @@
-var config = __config;
-
-module.exports = {
-    dep: [],
-    fn: function (gulp, callback) {
-        return gulp.src(
-                'annotations.js',
-                {
-                    cwd: config.paths.source.data
-                }
-            )
-            .pipe(gulp.dest(config.paths.public.data))
-        ;
-    }
-};
